+++ lib/presentation/app.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../localizations/strings.g.dart';
+import '../utils/localizations/strings.g.dart';
 import '../utils/logger.dart';
 import 'components/router.dart';
 import 'components/theme.dart';
