+++ test/presentation/pages/error/error_page_test.dart
@@ -4,8 +4,8 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/localizations/strings.g.dart';
 import 'package:github_search/presentation/pages/error/error_page.dart';
+import 'package:github_search/utils/localizations/strings.g.dart';
 
 import '../../../test_utils/golden_testing_tools.dart';
 import '../../../test_utils/test_agent.dart';
