+++ test/presentation/pages/repo/repo_index_page_test.dart
@@ -5,13 +5,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/infrastructure/github/http_client.dart';
-import 'package:github_search/localizations/strings.g.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_order.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_query.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_sort.dart';
 import 'package:github_search/presentation/pages/repo/repo_index_page.dart';
 import 'package:github_search/presentation/pages/repo/repo_search_page.dart';
 import 'package:github_search/presentation/pages/repo/repo_view_page.dart';
+import 'package:github_search/utils/localizations/strings.g.dart';
 
 import '../../../test_utils/logger.dart';
 import '../../../test_utils/mocks.dart';
