+++ test/presentation/pages/repo/repo_search_page_test.dart
@@ -6,11 +6,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/domain/repositories/query_history/entities/query_history_input.dart';
 import 'package:github_search/domain/repositories/query_history/query_history_repository.dart';
-import 'package:github_search/localizations/strings.g.dart';
 import 'package:github_search/presentation/pages/repo/components/query_histories.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_query.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_sort.dart';
 import 'package:github_search/presentation/pages/repo/repo_search_page.dart';
+import 'package:github_search/utils/localizations/strings.g.dart';
 
 import '../../../test_utils/golden_testing_tools.dart';
 import '../../../test_utils/test_agent.dart';
